// src/types/user.ts
export type IsoAutoDateString = string & { readonly autodate: unique symbol };
export type RecordIdString = string;
export type FileNameString = string & { readonly filename: unique symbol };
// ...
export enum UsersRoleOptions {
  agent = "agent",
  manager = "manager",
}
export type UsersRecord = {
  avatar?: FileNameString;
  created: IsoAutoDateString;
  email: string;
  emailVisibility?: boolean;
  externalId: string;
  id: string;
  name: string;
  password: string;
  phone: string;
  role: UsersRoleOptions;
  tokenKey: string;
  updated: IsoAutoDateString;
  verified?: boolean;
};